requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
